Reorder some stuff
