+++ src/features/kartleggingssporsmal/form/formDefaultValues.ts
@@ -1,0 +1,12 @@
+import type { KartleggingssporsmalForm } from "@/forms/kartleggingssporsmalForm";
+
+type KartleggingssporsmalFormAlsoUnfilled = {
+  [K in keyof KartleggingssporsmalForm]: KartleggingssporsmalForm[K] | "";
+};
+
+export const formDefaultValues: KartleggingssporsmalFormAlsoUnfilled = {
+  hvorSannsynligTilbakeTilJobben: "",
+  hvorSannsynligTilbakeTilJobbenBegrunnelse: "",
+  samarbeidOgRelasjonTilArbeidsgiver: "",
+  naarTilbakeTilJobben: "",
+};
